requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow-cpu==2.12.0
pillow==10.3.0
numpy==1.23.5
python-multipart==0.0.9
gdown==5.1.0
requests==2.31.0
